refactor(grocery,meal,aichat): default Content() for grocery + meal + aichat blocs
Move GroceryDetailScreen (which hosts GroceryDetailSheetContent),
EditGroceryListScreen, GroceryListScreen, MealPlanScreen, AiChatScreen,
AiChatHistoryScreen, and AiChatRootScreen into their public modules and
give each Bloc interface a default Content() that delegates to the screen.
Adds util.public + multiplatform-markdown deps to aichat/public so the
chat screen can compile in the public module. Impl Blocs and preview
fakes drop their Content overrides; screenshot tests now import the
screen composables from the new public packages (preview vals stay in
impl).
