+++ client/aichat/public/src/commonMain/kotlin/com/plusmobileapps/chefmate/aichat/AiChatBloc.kt
@@ -1,5 +1,7 @@
 package com.plusmobileapps.chefmate.aichat
 
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
 import com.plusmobileapps.chefmate.BackClickBloc
 import com.plusmobileapps.chefmate.BlocContext
 import com.plusmobileapps.chefmate.Consumer
@@ -13,6 +15,11 @@ import kotlinx.serialization.Serializable
 
 interface AiChatBloc : BackClickBloc, ComposeScreen {
     val state: StateFlow<Model>
+
+    @Composable
+    override fun Content(modifier: Modifier) {
+        AiChatScreen(bloc = this, modifier = modifier)
+    }
 
     val inputText: StateFlow<String>
