+++ client/grocery/core/impl/src/commonMain/kotlin/com/plusmobileapps/chefmate/grocery/core/impl/detail/GroceryDetailBlocImpl.kt
@@ -1,14 +1,11 @@
 package com.plusmobileapps.chefmate.grocery.core.impl.detail
 
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
 import com.plusmobileapps.chefmate.BlocContext
 import com.plusmobileapps.chefmate.Consumer
 import com.plusmobileapps.chefmate.di.AppScope
 import com.plusmobileapps.chefmate.getViewModel
 import com.plusmobileapps.chefmate.grocery.core.detail.GroceryDetailBloc
 import com.plusmobileapps.chefmate.grocery.core.detail.GroceryDetailBloc.Output
-import com.plusmobileapps.chefmate.grocery.core.impl.detail.ui.GroceryDetailSheetContent
 import com.plusmobileapps.chefmate.grocery.data.GroceryCategory
 import com.plusmobileapps.chefmate.mapState
 import com.plusmobileapps.metro.extensions.assistedfactory.ContributesAssistedFactory
@@ -74,10 +71,5 @@ class GroceryDetailBlocImpl(
 
     override fun onBackClicked() {
         output.onNext(Output.Finished)
-    }
-
-    @Composable
-    override fun Content(modifier: Modifier) {
-        GroceryDetailSheetContent(bloc = this, modifier = modifier)
     }
 }
