+++ client/grocery/core/public/src/commonMain/kotlin/com/plusmobileapps/chefmate/grocery/core/detail/GroceryDetailBloc.kt
@@ -1,5 +1,7 @@
 package com.plusmobileapps.chefmate.grocery.core.detail
 
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
 import com.plusmobileapps.chefmate.BackClickBloc
 import com.plusmobileapps.chefmate.BlocContext
 import com.plusmobileapps.chefmate.Consumer
@@ -10,6 +12,11 @@ import kotlinx.coroutines.flow.StateFlow
 
 interface GroceryDetailBloc : BackClickBloc, ComposeScreen {
     val models: StateFlow<Model>
+
+    @Composable
+    override fun Content(modifier: Modifier) {
+        GroceryDetailSheetContent(bloc = this, modifier = modifier)
+    }
 
     fun onGroceryNameChanged(name: String)
