+++ client/grocery/core/public/src/commonMain/kotlin/com/plusmobileapps/chefmate/grocery/core/detail/GroceryDetailScreen.kt
@@ -1,6 +1,6 @@
 @file:OptIn(ExperimentalMaterial3Api::class)
 
-package com.plusmobileapps.chefmate.grocery.core.impl.detail.ui
+package com.plusmobileapps.chefmate.grocery.core.detail
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
@@ -37,8 +37,6 @@ import chefmate.client.grocery.core.public.generated.resources.grocery_recipe_so
 import chefmate.client.grocery.core.public.generated.resources.purchased
 import chefmate.client.ui.public.generated.resources.Res as CommonRes
 import chefmate.client.ui.public.generated.resources.save
-import com.plusmobileapps.chefmate.grocery.core.detail.GroceryDetailBloc
-import com.plusmobileapps.chefmate.grocery.core.detail.GroceryDetailTestTags
 import com.plusmobileapps.chefmate.grocery.core.displayName
 import com.plusmobileapps.chefmate.grocery.data.GroceryCategory
 import com.plusmobileapps.chefmate.grocery.data.GroceryItem
