+++ client/grocery/core/public/src/commonMain/kotlin/com/plusmobileapps/chefmate/grocery/core/edit/EditGroceryListBloc.kt
@@ -1,5 +1,7 @@
 package com.plusmobileapps.chefmate.grocery.core.edit
 
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
 import com.plusmobileapps.chefmate.BackClickBloc
 import com.plusmobileapps.chefmate.BlocContext
 import com.plusmobileapps.chefmate.Consumer
@@ -11,6 +13,11 @@ import kotlinx.coroutines.flow.StateFlow
 
 interface EditGroceryListBloc : BackClickBloc, ComposeScreen {
     val state: StateFlow<Model>
+
+    @Composable
+    override fun Content(modifier: Modifier) {
+        EditGroceryListScreen(bloc = this, modifier = modifier)
+    }
 
     fun onNameChanged(name: String)
