+++ client/meal/core/public/src/commonMain/kotlin/com/plusmobileapps/chefmate/meal/core/MealPlanBloc.kt
@@ -1,5 +1,7 @@
 package com.plusmobileapps.chefmate.meal.core
 
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
 import com.plusmobileapps.chefmate.BlocContext
 import com.plusmobileapps.chefmate.Consumer
 import com.plusmobileapps.chefmate.meal.data.MealPlanItem
@@ -14,6 +16,11 @@ import kotlinx.datetime.LocalDate
 
 interface MealPlanBloc : ComposeScreen {
     val state: StateFlow<Model>
+
+    @Composable
+    override fun Content(modifier: Modifier) {
+        MealPlanScreen(bloc = this, modifier = modifier)
+    }
 
     fun onViewModeSelected(mode: ViewMode)
